move sim functions to new file + class
