+++ Scripts/s4ap/events/career_event_dispatcher.py
@@ -4,6 +4,7 @@ from s4ap.events.checks.send_check_event import SendLocationEvent
 from s4ap.logging.s4ap_logger import S4APLogger
 from s4ap.modinfo import ModInfo
 from s4ap.utils.s4ap_generic_utils import S4APUtils
+from s4ap.utils.s4ap_sim_utils import S4APSimUtils
 from sims4communitylib.events.event_handling.common_event import CommonEvent
 from sims4communitylib.events.event_handling.common_event_registry import CommonEventRegistry
 from sims4communitylib.services.common_service import CommonService
@@ -45,7 +46,7 @@ def _on_milestone_complete(original, self, *args, **kwargs):
     level = self._user_level
     sim_info = self._sim_info
     if sim_info in CommonHouseholdUtils.get_sim_info_of_all_sims_in_active_household_generator():
-        sim_name = S4APUtils.get_sim_first_name(sim_info)
+        sim_name = S4APSimUtils.get_sim_first_name(sim_info)
         OnCareerPromotionEvent.get()._on_promotion(career, level, sim_name)
     return result
