+++ Scripts/s4ap/utils/s4ap_generic_utils.py
@@ -30,14 +30,3 @@ class S4APUtils:
             CommonExceptionHandler.log_exception(ModInfo.get_identity(), 'An exception occurred while autosaving.',
                                                  exception=ex)
             return False
-
-
-    def get_sim_first_name(sim_info: SimInfo):
-        if sim_info is None or not hasattr('first_name'):
-            return ''
-        return getattr(sim_info, 'first_name')
-
-    @classmethod
-    def get_sim_instance(cls, sim_identifier: SimInfo):
-        if isinstance(sim_identifier, SimInfo):
-            return sim_identifier.get_sim_instance()
